Refactor curves and schemes into folders
